WeatherMatrix/requirements.txt
requests>=2.31.0
pytest>=7.4.0
pytest-mock>=3.11.0
python-dotenv>=1.0.0
Pillow>=10.0.0

